Add USB Notify and OTG Power Regulator to Misc
Tab 'Misc' will display now if the device does support Samsung's USB Notify in kernel. Additionally, the current state of the OTG Power Regulator will be displayed if setting 'Dislay more information' is enabled. The access to read the state of the OTG Power Regulator may be blocked by SELinux.